+++ aiworkspace-20210204/include/alibabacloud/models/ListImagesResponseBodyImages.hpp
@@ -129,8 +129,8 @@ namespace Models
     // size Field Functions 
     bool hasSize() const { return this->size_ != nullptr;};
     void deleteSize() { this->size_ = nullptr;};
-    inline int32_t size() const { DARABONBA_PTR_GET_DEFAULT(size_, 0) };
-    inline ListImagesResponseBodyImages& setSize(int32_t size) { DARABONBA_PTR_SET_VALUE(size_, size) };
+    inline int64_t size() const { DARABONBA_PTR_GET_DEFAULT(size_, 0L) };
+    inline ListImagesResponseBodyImages& setSize(int64_t size) { DARABONBA_PTR_SET_VALUE(size_, size) };
 
 
     // sourceId Field Functions 
@@ -184,7 +184,7 @@ namespace Models
     // The ID of the Alibaba Cloud account.
     std::shared_ptr<string> parentUserId_ = nullptr;
     // The image size. Unit: GB.
-    std::shared_ptr<int32_t> size_ = nullptr;
+    std::shared_ptr<int64_t> size_ = nullptr;
     // 镜像来源 ID
     std::shared_ptr<string> sourceId_ = nullptr;
     // 镜像来源类型
